fix: fix playing music and update full player screen

+++ core/model/src/main/java/com/donfreddy/troona/core/model/data/UserData.kt
@@ -38,6 +38,6 @@ data class UserData(
   val useDynamicColor: Boolean,
   val playbackMode: PlaybackMode,
   val playingQueueIndex: Int,
-  val playingQueuePosition: Int,
+  val playingQueuePosition: Long,
   val playingQueueIds: List<String>,
 )
